// frontend/src/client/types.gen.ts
export type ArticleRead = {
    /**
     * Id
     */
    id: string;
    /**
     * Display Name
     */
    display_name: string;
    /**
     * Text
     */
    text: string;
    /**
     * Creator Id
     */
    creator_id: string;
    /**
     * Created At
     */
    created_at: string;
    /**
     * Updated At
     */
    updated_at: string;
    primary_image: ImageRead | null;
    background_image: ImageRead | null;
};
// ...
export type ImageRead = {
    /**
     * Id
     */
    id: string;
    /**
     * Display Name
     */
    display_name: string;
    /**
     * Content Type
     */
    content_type: string;
};

/**
 * UserCreate
 */
export type UserCreate = {
    /**
     * Username
     */
    username: string;
    /**
     * Display Name
     */
    display_name: string;
    /**
     * Password
     */
    password: string;
};

/**
 * UserRead
 */
export type UserRead = {
    /**
     * Id
     */
    id: string;
    /**
     * Username
     */
    username: string;
    /**
     * Display Name
     */
    display_name: string;
    /**
     * Created At
     */
    created_at: string;
    /**
     * Updated At
     */
    updated_at: string;
    avatar: ImageRead | null;
};